pra 98 feat: вынесена в отдельный компонент кнопка, отвечающая за создание темы
