+++ packages/client/src/pages/Forum/components/CommentsPage/CommentsPage.module.scss
@@ -5,7 +5,3 @@
   flex-flow: column;
   height: 500px;
 }
-
-.button_new_post {
-  margin: 10px 50px 40px;
-}
