+++ packages/client/src/pages/Forum/components/CommentsPage/CommentsPage.tsx
@@ -1,30 +1,18 @@
-import React, { useCallback } from 'react';
+import React from 'react';
 import styles from './CommentsPage.module.scss';
-import { Button } from 'src/components/Button';
-import { useNavigate } from 'react-router-dom';
 import { ErrorBoundary } from 'src/components/ErrorBoundary';
 import { useAppSelector } from 'src/hooks/redux';
 import { selectActiveTopic } from 'src/store/forum/ForumSelectors';
 import { BlockComments } from '../../part/BlockComments';
+import { ButtonCreateTopic } from '../../part/ButtonCreateTopic';
 
 export const CommentsPage = () => {
-  const navigate = useNavigate();
-
   const [topic] = useAppSelector(selectActiveTopic);
-
-  const handleCreateTopic = useCallback(() => {
-    navigate('/forum/createtopic');
-  }, [navigate]);
 
   return (
     <ErrorBoundary>
       <div className={styles.block}>
-        <Button
-          regular
-          className={styles.button_new_post}
-          onClick={handleCreateTopic}>
-          POST NEW TOPIC
-        </Button>
+        <ButtonCreateTopic />
         {topic && <BlockComments topic={topic} />}
       </div>
     </ErrorBoundary>
