+++ packages/client/src/pages/Forum/part/ButtonCreateTopic/ButtonCreateTopic.module.scss
@@ -1,0 +1,4 @@
+.button_create_topic {
+    margin: 10px 50px 40px;
+  }
+
