+++ packages/client/src/pages/Forum/part/ButtonCreateTopic/index.ts
@@ -1,0 +1,1 @@
+export { ButtonCreateTopic } from './ButtonCreateTopic';
